fix: 변수 타입 변경

+++ src/main/java/com/letsparty/vo/Media.java
@@ -19,6 +19,6 @@ public class Media {
 	private boolean isDeleted;
 	private String name;
 	private String path;
-	private Long size;
-	private long postId;
+	private Integer size;
+	private Long postId;
 }
